requirements.txt
# CrewAI Moving Services Metadata Generation Pipeline Requirements

# Core CrewAI and AI frameworks
crewai>=0.74.0
crewai-tools>=0.12.0

# AWS SDK for Athena and Glue integration
boto3>=1.34.0
botocore>=1.34.0

# Weaviate client for vector database
weaviate-client>=4.9.0

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0

# YAML processing
pyyaml>=6.0

# Environment management
python-dotenv>=1.0.0

# Logging and utilities
colorlog>=6.7.0

# HTTP requests (for API calls)
requests>=2.31.0

# JSON handling improvements
ujson>=5.8.0

# Date/time handling
python-dateutil>=2.8.2

# Statistical analysis
scipy>=1.11.0

# Type hints support
typing-extensions>=4.8.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0